+++ src/MySqlConnector/Protocol/Serialization/BufferedByteReader.cs
@@ -5,6 +5,11 @@ namespace MySql.Data.Protocol.Serialization
 {
 	internal sealed class BufferedByteReader
 	{
+		public BufferedByteReader()
+		{
+			m_buffer = new byte[16384];
+		}
+
 		public ValueTask<ArraySegment<byte>> ReadBytesAsync(IByteHandler byteHandler, int count, IOBehavior ioBehavior)
 		{
 			if (m_remainingData.Count >= count)
@@ -14,50 +19,42 @@ namespace MySql.Data.Protocol.Serialization
 				return new ValueTask<ArraySegment<byte>>(readBytes);
 			}
 
-			if (m_remainingData.Count == 0)
-				return ReadBytesAsync(byteHandler, default(ArraySegment<byte>), count, ioBehavior);
+			var buffer = count > m_buffer.Length ? new byte[count] : m_buffer;
+			if (m_remainingData.Count > 0)
+			{
+				Buffer.BlockCopy(m_remainingData.Array, m_remainingData.Offset, buffer, 0, m_remainingData.Count);
+				m_remainingData = new ArraySegment<byte>(buffer, 0, m_remainingData.Count);
+			}
 
-			// save data from m_remainingData.Array because calling ReadAsync may invalidate it
-			var buffer = new byte[Math.Max(count, 16384)];
-			Buffer.BlockCopy(m_remainingData.Array, m_remainingData.Offset, buffer, 0, m_remainingData.Count);
-			var previousReadBytes = new ArraySegment<byte>(buffer, 0, m_remainingData.Count);
-
-			return ReadBytesAsync(byteHandler, previousReadBytes, count, ioBehavior);
+			return ReadBytesAsync(byteHandler, new ArraySegment<byte>(buffer, m_remainingData.Count, buffer.Length - m_remainingData.Count), count, ioBehavior);
 		}
 
-		private ValueTask<ArraySegment<byte>> ReadBytesAsync(IByteHandler byteHandler, ArraySegment<byte> previousReadBytes, int count, IOBehavior ioBehavior)
+		private ValueTask<ArraySegment<byte>> ReadBytesAsync(IByteHandler byteHandler, ArraySegment<byte> buffer, int count, IOBehavior ioBehavior)
 		{
-			return byteHandler.ReadBytesAsync(count - previousReadBytes.Count, ioBehavior)
-				.ContinueWith(readBytes =>
+			return byteHandler.ReadBytesAsync(buffer, ioBehavior)
+				.ContinueWith(readBytesCount =>
 				{
-					if (readBytes.Count == 0)
-						return new ValueTask<ArraySegment<byte>>(previousReadBytes);
-
-					if (previousReadBytes.Array == null && readBytes.Count >= count)
+					if (readBytesCount == 0)
 					{
-						m_remainingData = readBytes.Slice(count);
-						return new ValueTask<ArraySegment<byte>>(readBytes.Slice(0, count));
+						var data = m_remainingData;
+						m_remainingData = default(ArraySegment<byte>);
+						return new ValueTask<ArraySegment<byte>>(data);
 					}
 
-					var previousReadBytesArray = previousReadBytes.Array;
-					if (previousReadBytesArray == null)
-						previousReadBytesArray = new byte[Math.Max(count, 16384)];
-					else if (previousReadBytesArray.Length < previousReadBytes.Count + readBytes.Count)
-						Array.Resize(ref previousReadBytesArray, Math.Max(previousReadBytesArray.Length * 2, previousReadBytes.Count + readBytes.Count));
-
-					Buffer.BlockCopy(readBytes.Array, readBytes.Offset, previousReadBytesArray, previousReadBytes.Offset + previousReadBytes.Count, readBytes.Count);
-					previousReadBytes = new ArraySegment<byte>(previousReadBytesArray, previousReadBytes.Offset, previousReadBytes.Count + readBytes.Count);
-
-					if (previousReadBytes.Count >= count)
+					var bufferSize = buffer.Offset + readBytesCount;
+					if (bufferSize >= count)
 					{
-						m_remainingData = previousReadBytes.Slice(count);
-						return new ValueTask<ArraySegment<byte>>(previousReadBytes.Slice(0, count));
+						var bufferBytes = new ArraySegment<byte>(buffer.Array, 0, bufferSize);
+						var requestedBytes = bufferBytes.Slice(0, count);
+						m_remainingData = bufferBytes.Slice(count);
+						return new ValueTask<ArraySegment<byte>>(requestedBytes);
 					}
 
-					return ReadBytesAsync(byteHandler, previousReadBytes, count, ioBehavior);
+					return ReadBytesAsync(byteHandler, new ArraySegment<byte>(buffer.Array, bufferSize, buffer.Array.Length - bufferSize), count, ioBehavior);
 				});
 		}
 
 		ArraySegment<byte> m_remainingData;
+		readonly byte[] m_buffer;
 	}
 }