+++ src/MySqlConnector/Utility.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Globalization;
 using System.IO;
+using System.Runtime.InteropServices;
 using System.Text;
 
 namespace MySql.Data
@@ -94,6 +95,19 @@ namespace MySql.Data
 					writer.Write((byte) (0x80 | (codePoint & 0x3F)));
 					startIndex++;
 				}
+			}
+		}
+
+		public static bool IsWindows()
+		{
+			try
+			{
+				// OSPlatform.Windows is not supported on AWS Lambda
+				return RuntimeInformation.IsOSPlatform(OSPlatform.Windows);
+			}
+			catch (PlatformNotSupportedException)
+			{
+				return false;
 			}
 		}
 	}