+++ tests/MySqlConnector.Performance/Controllers/SyncController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using System;
+using Microsoft.AspNetCore.Mvc;
 using MySqlConnector.Performance.Models;
 
 namespace MySqlConnector.Performance.Controllers
@@ -6,7 +7,7 @@ namespace MySqlConnector.Performance.Controllers
 	[Route("api/[controller]")]
 	public class SyncController : Controller
 	{
-		// GET api/async
+		// GET api/sync
 		[HttpGet]
 		public IActionResult GetLatest()
 		{
@@ -19,7 +20,7 @@ namespace MySqlConnector.Performance.Controllers
 			}
 		}
 
-		// GET api/async/5
+		// GET api/sync/5
 		[HttpGet("{id}")]
 		public IActionResult GetOne(int id)
 		{
@@ -34,7 +35,7 @@ namespace MySqlConnector.Performance.Controllers
 			}
 		}
 
-		// POST api/async
+		// POST api/sync
 		[HttpPost]
 		public IActionResult Post([FromBody]BlogPost body)
 		{
@@ -47,7 +48,7 @@ namespace MySqlConnector.Performance.Controllers
 			}
 		}
 
-		// PUT api/async/5
+		// PUT api/sync/5
 		[HttpPut("{id}")]
 		public IActionResult PutOne(int id, [FromBody]BlogPost body)
 		{
@@ -65,7 +66,7 @@ namespace MySqlConnector.Performance.Controllers
 			}
 		}
 
-		// DELETE api/async/5
+		// DELETE api/sync/5
 		[HttpDelete("{id}")]
 		public IActionResult DeleteOne(int id)
 		{
@@ -81,7 +82,7 @@ namespace MySqlConnector.Performance.Controllers
 			}
 		}
 
-		// DELETE api/async
+		// DELETE api/sync
 		[HttpDelete]
 		public IActionResult DeleteAll()
 		{
@@ -91,6 +92,78 @@ namespace MySqlConnector.Performance.Controllers
 				var query = new BlogPostQuery(db);
 				query.DeleteAll();
 				return new OkResult();
+			}
+		}
+
+		// GET api/sync/hello
+		// This method is used to establish baseline web server performance
+		// i.e. how many base RPS can the server handle
+		[HttpGet("hello")]
+		public IActionResult Hello()
+		{
+			return new OkObjectResult("hello world");
+		}
+
+		// GET api/sync/bulkinsert/10000
+		[HttpGet("bulkinsert/{num}")]
+		public IActionResult BulkInsert(int num)
+		{
+			using (var db = new AppDb())
+			{
+				var time = DateTime.Now;
+				db.Connection.Open();
+				var txn = db.Connection.BeginTransaction();
+				try
+				{
+					for (var i = 0; i < num; i++)
+					{
+						var blogPost = new BlogPost
+						{
+							Db = db,
+							Title = "bulk",
+							Content = "bulk " + num
+						};
+						blogPost.Insert();
+					}
+				}
+				catch (Exception)
+				{
+					txn.Rollback();
+					throw;
+				}
+				txn.Commit();
+				var timing = $"Sync: Inserted {num} records in " + (DateTime.Now - time);
+				Console.WriteLine(timing);
+				return new OkObjectResult(timing);
+			}
+		}
+
+		// GET api/sync/bulkselect
+		[HttpGet("bulkselect/{num}")]
+		public IActionResult BulkSelect(int num)
+		{
+			using (var db = new AppDb())
+			{
+				var time = DateTime.Now;
+				db.Connection.Open();
+				var query = new BlogPostQuery(db);
+				var reader = query.LatestPostsCmd(num).ExecuteReader();
+
+				var numRead = 0;
+				while (reader.Read())
+				{
+					var post = new BlogPost(db)
+					{
+						Id = reader.GetFieldValue<int>(0),
+						Title = reader.GetFieldValue<string>(1),
+						Content = reader.GetFieldValue<string>(2)
+					};
+					numRead++;
+				}
+
+				var timing = $"Sync: Read {numRead} records in " + (DateTime.Now - time);
+				Console.WriteLine(timing);
+				return new OkObjectResult(timing);
 			}
 		}
 	}