+++ tests/SideBySide.New/ConnectSync.cs
@@ -9,7 +9,6 @@ using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
 using MySql.Data.MySqlClient;
-using SideBySide.New;
 using Xunit;
 
 namespace SideBySide