+++ tests/SideBySide.New/StoredProcedureTests.cs
@@ -3,7 +3,6 @@ using System.Data;
 using System.Data.Common;
 using System.Threading.Tasks;
 using MySql.Data.MySqlClient;
-using SideBySide.New;
 using Xunit;
 
 namespace SideBySide