+++ docs/VersionHistory.md
@@ -12,6 +12,26 @@ of supported options.
 
 ## Release Notes
 
+### 0.12.0
+
+* Add support for `DateTimeOffset`: [#172](https://github.com/mysql-net/MySqlConnector/issues/172), [#175](https://github.com/mysql-net/MySqlConnector/issues/175).
+
+### 0.11.6
+
+* Fix `PlatformNotSupportedException` on AWS Lambda: [#170](https://github.com/mysql-net/MySqlConnector/issues/170).
+
+### 0.11.5
+
+* Further improve async and sync performance: [#164](https://github.com/mysql-net/MySqlConnector/issues/164).
+
+### 0.11.4
+
+* No changes in this release.
+
+### 0.11.3
+
+* Improve async performance: [#164](https://github.com/mysql-net/MySqlConnector/issues/164).
+
 ### 0.11.2
 
 * Fix InvalidCastException when using aggregate functions: [#54](https://github.com/mysql-net/MySqlConnector/issues/54).